backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
supabase==2.13.0
google-generativeai==0.8.4
celery[redis]==5.4.0
redis==5.2.1
python-multipart==0.0.20
pypdf==5.1.0
python-docx==1.1.2
pydantic-settings==2.7.1
PyJWT==2.10.1
httpx==0.28.1
